LottoGenerator 인터페이스화
