+++ src/main/java/kr/codesquad/Model/LottoGenerator.java
@@ -1,33 +1,11 @@
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by FernFlower decompiler)
-//
-
 package kr.codesquad.Model;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class LottoGenerator {
-    public LottoGenerator() {
-    }
+public interface LottoGenerator {
+    Lotto getNewLotto();
+    List<Integer> setLottoNum(List<Integer> lottoNum);
 
-    public Lotto getNewLotto() {
-        List<Integer> lottoNum = Lotto.lottoNum;
-        Collections.shuffle(lottoNum);
-        List<Integer> lottoNumList = this.setLottoNum(lottoNum);
-        Lotto lotto = new Lotto(lottoNumList);
-        return lotto;
-    }
-
-    private List<Integer> setLottoNum(List<Integer> lottoNum) {
-        List<Integer> number = new ArrayList();
-
-        for(int i = 0; i < 6; ++i) {
-            number.add((Integer)lottoNum.get(i));
-        }
-
-        return number;
-    }
 }
